+++ tests/bchain/posts_test.py
@@ -4,6 +4,7 @@ from tests.fixtures import random_data
 
 import trufflepig.bchain.posts as tbpo
 import trufflepig.preprocessing as tppp
+import trufflepig.filters.textfilters as tptf
 
 
 def test_comment():
@@ -23,14 +24,16 @@ def test_topN_post():
     df = tppp.preprocess(df, ncores=1)
 
     date = pd.datetime.utcnow().date()
+    df.image_urls = df.body.apply(lambda x: tptf.get_image_urls(x))
 
     title, post = tbpo.topN_post(topN_authors=df.author,
-                             topN_permalinks=df.permalink,
-                             topN_titles=df.title,
-                             topN_filtered_bodies=df.filtered_body,
-                             topN_votes=df.votes,
-                             topN_rewards=df.reward,
-                             title_date=date)
+                                 topN_permalinks=df.permalink,
+                                 topN_titles=df.title,
+                                 topN_filtered_bodies=df.filtered_body,
+                                 topN_image_urls=df.image_urls,
+                                 topN_votes=df.votes,
+                                 topN_rewards=df.reward,
+                                 title_date=date)
 
     assert post
     assert title