+++ trufflepig/filters/textfilters.py
@@ -17,6 +17,18 @@ def filter_images_and_links(text):
     return text
 
 
+def get_image_urls(text):
+    images = re.findall('!\[[-a-zA-Z0-9?@: %._\+~#=/()]*\]\([-a-zA-Z0-9?@:%._\+~#=/()]+\)|<img[^>]+src="[^">]+"[^>]*>', text)
+    image_urls = []
+    for image in images:
+        if image.startswith('<img'):
+            image_url = re.sub('<img[^>]+src="([^">]+)"[^>]*>',  '\g<1>', image)
+        else:
+            image_url = re.sub('!\[[-a-zA-Z0-9?@: %._\+~#=/()]*\]\(([-a-zA-Z0-9?@:%._\+~#=/()]+)\)', '\g<1>', image)
+        image_urls.append(image_url)
+    return image_urls
+
+
 def filter_urls(text):
     return re.sub('(https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9]'
                    '[a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|www\.'
@@ -47,7 +59,7 @@ def filter_punctuation(text):
 
 
 def filter_headings(text):
-    return re.sub('(\n|^)(#{1,6}| *<h[1-6]>)[^\n#]*\n', '', text)
+    return re.sub('(\n|^)( *#{1,6}| *<h[1-6]>)[^\n#]*\n', '', text)
 
 
 def is_in_filter_tags(tags, filter_tags):